Moved event card component to events folder
