+++ frontend-site/src/components/Events/EventCard.vue
