+++ frontend-site/src/views/Events.vue
@@ -4,7 +4,7 @@
 
 <script>
 import { mapState, mapActions } from 'vuex';
-import EventCard from '@/components/Home/EventCard';
+import EventCard from '@/components/Events/EventCard';
 
 export default {
   components: {
